requirements.txt
discord.py
fastapi
uvicorn
gTTS
python-dotenv
pydantic
pynacl
aiohttp
